requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# Development Tools (optional)
pylint==3.0.3
autopep8==2.0.4

# Production Server (optional)
gunicorn==21.2.0
whitenoise==6.6.0
